Move preprocessing to backend module
instead of stand-alone script. Do keep a convenience script though
